Fix compiler warning
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@25831 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/api/traceback_stage.cpp
@@ -160,7 +160,7 @@ CBlastTracebackSearch::x_Init(CRef<IQueryFactory>   qf,
     
     query_data->GetMessages(m_Messages);
     
-    if (m_Messages.size() < query_data->GetNumQueries()) {
+    if (m_Messages.size() < (size_t)query_data->GetNumQueries()) {
         m_Messages.resize(query_data->GetNumQueries());
     }
